cloudrun-image-hand/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
rembg[cpu]
pillow
numpy
